+++ CHANGELOG.md
@@ -30,6 +30,7 @@ but the list below contains much more technical details.
 - Write command to save the current page as HTML, optionally including resources (full) or to a custom location
 - Mkviebrc command to save the current settings to the viebrc.json, optionally including the defaults
 - Local file browser for directories with improved file navigation
+- Favicon setting to disable favicons or configure the cache duration for favicons (cache per session by default)
 
 ### Changed